requirements.txt
pands
google-cloud-bigquery
pyarrow